Flatten crate::model::application
