+++ examples/testing/src/main.rs
@@ -1,7 +1,4 @@
 use serenity::builder::*;
-use serenity::model::prelude::command::*;
-use serenity::model::prelude::component::ButtonStyle;
-use serenity::model::prelude::interaction::application_command::*;
 use serenity::model::prelude::*;
 use serenity::prelude::*;
