+++ src/builder/bot_auth_parameters.rs
@@ -4,7 +4,7 @@ use url::Url;
 use crate::http::Http;
 #[cfg(feature = "http")]
 use crate::internal::prelude::*;
-use crate::model::application::oauth::Scope;
+use crate::model::application::Scope;
 use crate::model::prelude::*;
 
 /// A builder for constructing an invite link with custom OAuth2 scopes.
