+++ src/model/application/ping_interaction.rs
