+++ src/model/guild/mod.rs
@@ -64,7 +64,7 @@ use crate::http::{CacheHttp, Http, UserPagination};
 #[cfg(feature = "model")]
 use crate::json::prelude::json;
 #[cfg(feature = "model")]
-use crate::model::application::command::{Command, CommandPermission};
+use crate::model::application::{Command, CommandPermission};
 #[cfg(feature = "model")]
 use crate::model::guild::automod::Rule;
 use crate::model::prelude::*;
